# perf: Improve performance when reading dynamic properties.