GitHub Issue: #
<!-- Link to relevant GitHub issue if applicable.
     All PRs should be associated with an issue -->

## Proposed Changes
<!-- Please check one or more that apply to this PR. -->

 - [ ] Bug fix
 - [ ] Feature
 - [ ] Code style update (formatting)
 - [ ] Refactoring (no functional changes, no api changes)
 - [ ] Build or CI related changes
 - [ ] Documentation content changes
 - [x] Other, please describe: Performance optimizations

## What is the current behavior?
<!-- Please describe the current behavior that you are modifying,
     or link to a relevant issue. -->

| Method                           | Mean        | Gen0   | Gen1   | Allocated |
|--------------------------------- |------------:|-------:|-------:|----------:|
| GetFromValue_Unresolved          |   427.69 ns | 0.0405 | 0.0401 |     256 B |
| GetFromObservable_Unresolved     |   544.88 ns | 0.0477 | 0.0467 |     304 B |
| GetFromValue_Resolved            |    46.09 ns | 0.0191 |      - |     120 B |
| GetFromObservable_Resolved       |    52.72 ns | 0.0204 |      - |     128 B |

## What is the new behavior?
<!-- Please describe the new behavior after your modifications. -->

- Faster reads and less memory allocations.
- No allocations when reading resolved properties.

| Method                           | Mean        | Gen0   | Gen1   | Allocated |
|--------------------------------- |------------:|-------:|-------:|----------:|
| GetFromValue_Unresolved          |   317.01 ns | 0.0253 | 0.0248 |     160 B |
| GetFromObservable_Unresolved     |   343.79 ns | 0.0315 | 0.0310 |     200 B |
| GetFromValue_Resolved            |    23.23 ns |      - |      - |         - |
| GetFromObservable_Resolved       |    30.41 ns |      - |      - |         - |

## Impact on version
<!-- Please select one or more based on your commits. -->

- [ ] **Major** (Public API was modified.)
  - Public constructs (class, struct, delegate, enum, etc.) were removed or renamed.
  - Public members were removed or renamed.
  - Public method signatures were changed or renamed.
- [ ] **Minor** (Public API was extended.)
  - Public constructs, members, or overloads were added.
- [x] **Patch** (Public API was unchanged.)
  - A bug in behavior was fixed.
  - Documentation was changed.
- [ ] **None** (The library is unchanged.)
  - Only code under the `build` folder was changed.
  - Only code under the `.github` folder was changed.
  - Only code in the Benchmarks project was changed.

## Checklist

Please check that your PR fulfills the following requirements:

- [x] Documentation has been added/updated.
- [ ] Automated Unit / Integration tests for the changes have been added/updated.
- [ ] Updated [BREAKING_CHANGES.md](../BREAKING_CHANGES.md) (if you introduced a breaking change).
- [x] Your conventional commits are aligned with the **Impact on version** section.

<!-- If this PR contains a breaking change, please describe the impact
     and migration path for existing applications below. -->

## Other information
<!-- Please provide any additional information if necessary -->